Use `botocore.UNSIGNED` signature version with custom config
The "signature_version" argument of `botocore.config.Config` accepts strings for signers and the `botocore.UNSIGNED` singleton when a signed should not be configured.

Change allows users to pass the `AWSCredentials` configuration {"signature_version": "unsigned"} and `prefect_aws.ClientParameters.get_params_override` will convert it to the singleton.

This should probably be handled within `botocore`, but it's a quick patch for accessing some [public buckets](https://registry.opendata.aws/noaa-ghcn/).